test: compare commit paths for algorithmic equality and dump replay
edge_locality_tests:
- test_weave_paths_algorithmic_mathematical_equality builds the same weighted
  graph through commit() and commit_unclustered(), then asserts PageRank scores
  agree within 1e-12 per node, that weakly connected components match exactly,
  and that K-hop node sets agree. Comparing the two commit paths directly is
  stronger than the existing per-edge-transaction comparison.
- test_mixed_transaction_never_takes_batch_path covers a transaction that mixes
  AddNode, AddEdge, RemoveEdge and UpdateNodeProperty, asserting the final graph
  is correct and the remaining edge is woven into both adjacency chains. Such a
  transaction cannot reach the batch path because weaving only merges
  consecutive AddEdge runs.

cypher_advanced_tests:
- test_dump_cypher_script_replay filters dump comments, splits the script on
  semicolons and replays every statement into a fresh in-memory database,
  asserting node, edge and property recovery. The existing CLI round trip
  replays line by line, so the script-level path was previously uncovered.

Also fixes an unnecessary u64 cast flagged by clippy in the density test.

+++ tests/slotted_property_tests.rs
@@ -300,7 +300,7 @@ fn test_forty_thousand_entity_density() -> Result<(), GraphError> {
 
     let num_nodes: u64 = 10_000;
     let num_edges: usize = 30_000;
-    let entities = num_nodes as u64 + num_edges as u64;
+    let entities = num_nodes + num_edges as u64;
     assert_eq!(entities, 40_000);
 
     {
